+++ akshare/__init__.py
@@ -1985,9 +1985,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.5.63 fix: fix import path problem
 1.5.64 fix: fix stock_cash_flow_sheet_by_yearly_em interface
 1.5.65 fix: fix stock_repurchase_em interface
+1.5.66 fix: fix crypto_hist interface
 """
 
-__version__ = "1.5.65"
+__version__ = "1.5.66"
 __author__ = "AKFamily"
 
 import sys